# check SPV proof inner nodes not to be valid transactions