This attempts to implement protection against the attack recently described [on the mailing list](https://lists.linuxfoundation.org/pipermail/bitcoin-dev/2018-June/016105.html).

Before merging, I would like to measure the performance impact with a relatively large wallet on mainnet.
So far, I could only measure on testnet, but blocks are mostly empty there, so merkle proofs are short.

On testnet, I synced a wallet (A) that has ~350 txns.
Irrespective of this PR, the sync took ~10 seconds.
Without the patch, the verifier spent about `19 ms` altogether in `hash_merkle_root`.
With the patch, the verifier spent about `66 ms` altogether in `hash_merkle_root`.

On testnet, I synced a wallet (B) that has ~10.9k txns.
The sync took ~3.5 minutes. (with the patch, it was about 4 seconds longer, but that is within error probably; I only measured this once)
Without the patch, the verifier spent about `733 ms` altogether in `hash_merkle_root`.
With the patch, the verifier spent about `2014 ms` altogether in `hash_merkle_root`.