@SomberNight 2018-06-15
I've found this address on mainnet: 1NcXyC1m3XE1s3u9rBZawt3sCZkeGEgoH9,
which has made 9733 transactions, all in the last 6 weeks.

When syncing a wallet for the above address, I benchmarked the verifier:
```
Without the patch, the verifier spent about `647 ms` altogether in `hash_merkle_root`.
With the patch, the verifier spent about `1946 ms` altogether in `hash_merkle_root`.
```